+++ code/src/components/App.css
@@ -1,0 +1,4 @@
+.content-container {
+  display: flex;
+  flex-direction: column;
+}
