+++ code/src/components/App.js
@@ -1,14 +1,15 @@
 import React from 'react'
 import { Header } from './Header'
-import { Questions } from './Questions'
+import { Form } from './Form'
+import './App.css'
 
 export const App = () => {
   return (
-    <>
+    <section className="content-container">
       <header>
         <Header />
       </header>
-      <Questions />
-    </>
+      <Form />
+    </section>
   )
 }
