+++ code/src/components/Form.css
@@ -1,0 +1,4 @@
+img {
+object-fit: cover;
+width: 200px;
+}
