+++ code/src/components/InputCheckbox.js
@@ -1,18 +1,18 @@
-import React, { useState } from 'react'
+// import React from 'react'
 
-export const InputCheckbox = (props) => {
-  //const [agreeOnTerms, setAgreeOnTerms] = useState(false)
-  const { question, checked, setAgreeOnTerms } = props
-  return (
-    <label>
-      <h2>{question}</h2>
-      <input
-        //key={agreeOnTerms.id}
-        type='checkbox'
-        //checked={agreeOnTerms}
-        checked={checked}
-        onChange={event => setAgreeOnTerms(event.target.checked)}
-      />
-    </label>
-  )
-}
+// export const InputCheckbox = (props) => {
+//   //const [agreeOnTerms, setAgreeOnTerms] = useState(false)
+//   const { question, checked, setAgreeOnTerms } = props
+//   return (
+//     <label>
+//       <h2>{question}</h2>
+//       <input
+//         //key={agreeOnTerms.id}
+//         type='checkbox'
+//         //checked={agreeOnTerms}
+//         checked={checked}
+//         onChange={event => setAgreeOnTerms(event.target.checked)}
+//       />
+//     </label>
+//   )
+// }
