+++ code/src/components/InputText.js
@@ -7,11 +7,10 @@ export const InputText = ({ id, question, value, setText }) => {
       <input
         type="text"
         id={id}
-        className="text-input"
         value={value}
         onChange={(event) => setText(event.target.value)}
-        //placeholder={placeholder}
-        onKeyPress={(event) => { event.key === 'Enter' && event.preventDefault(); }}
+        placeholder='Type your name'
+      //onKeyPress={(event) => { event.key === 'Enter' && event.preventDefault(); }}
 
       />
     </label>
