+++ code/src/components/InputTextTest.js
@@ -1,0 +1,15 @@
+import React, { useState } from 'react'
+
+export const InputTextTest = () => {
+  const [lastName, setLastName] = useState('');
+  return (
+    <label>
+      <h2>What is your last name?</h2>
+      <input
+        type='text'
+        onchange={event => setLastName(event.target.value)}
+        value={lastName}
+      />
+    </label>
+  )
+}
