--- code/src/components/Questions.css
@@ -1,4 +1,0 @@
-form {
-  display: flex;
-  flex-direction: column;
-}
