+++ code/src/components/StartButton.js
@@ -1,8 +1,8 @@
-import React from 'react'
+// import React from 'react'
 
-export const StartButton = ({ button, click, disable, text }) => {
-  return (
+// export const StartButton = ({ button, click, disable, text }) => {
+//   return (
 
-    <button type={button} onClick={click} disable={disable}>{text}</button>
-  )
-}
+//     <button type={button} onClick={click} disable={disable}>{text}</button>
+//   )
+// }
